+++ DangProject/Presentation Layer/Login/View/LoginViewController.swift
@@ -6,7 +6,6 @@
 //
 
 import AuthenticationServices
-import CryptoKit
 import Foundation
 import UIKit
 
@@ -14,13 +13,11 @@ import RxSwift
 import RxCocoa
 
 class LoginViewController: UIViewController {
-    
+    private let disposBag = DisposeBag()
     weak var coordinator: LoginCoordinator?
-    let viewModel: LoginViewModel
-    fileprivate var currentNonce: String?
-    var coordinatorFinishDelegate: CoordinatorFinishDelegate?
+    private let viewModel: LoginViewModel
     
-    let disposBag = DisposeBag()
+    var coordinatorFinishDelegate: CoordinatorFinishDelegate?
     // MARK: - Init
     init(viewModel: LoginViewModel) {
         self.viewModel = viewModel
@@ -57,75 +54,19 @@ class LoginViewController: UIViewController {
     
     private func bindSignInObservable() {
         viewModel.profileExistenceObservable
-            .bind(onNext: { [unowned self] profileIsValid in
+            .bind(onNext: { [weak self] profileIsValid in
                 if profileIsValid {
-                    coordinatorFinishDelegate?.switchViewController(to: .tabBar)
+                    self?.coordinatorFinishDelegate?.switchViewController(to: .tabBar)
                 } else {
-                    coordinatorFinishDelegate?.switchViewController(to: .inputPersonalInformation)
+                    self?.coordinatorFinishDelegate?.switchViewController(to: .inputPersonalInformation)
                 }
             })
             .disposed(by: disposBag)
     }
     
-    
     @objc func handleAuthorizationAppleIDButtonPress() {
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
+        viewModel.loginButtonDidTap(with: self)
     }
-    
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        var result = ""
-        var remainingLength = length
-        
-        while remainingLength > 0 {
-            let randoms: [UInt8] = (0 ..< 16).map { _ in
-                var random: UInt8 = 0
-                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-                if errorCode != errSecSuccess {
-                    fatalError(
-                        "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-                    )
-                }
-                return random
-            }
-            randoms.forEach { random in
-                if remainingLength == 0 {
-                    return
-                }
-                if random < charset.count {
-                    result.append(charset[Int(random)])
-                    remainingLength -= 1
-                }
-            }
-        }
-        
-        return result
-    }
-    
-    @available(iOS 13, *)
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
-    }
-    
-    
     
 }
 extension LoginViewController: ASAuthorizationControllerDelegate, ASAuthorizationControllerPresentationContextProviding {
@@ -135,24 +76,10 @@ extension LoginViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
     }
     
     func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
-        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-            guard let nonce = currentNonce else {
-                fatalError("Invalid state: A login callback was received, but no login request was sent.")
-            }
-            guard let appleIDToken = appleIDCredential.identityToken else {
-                print("Unable to fetch identity token")
-                return
-            }
-            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
-                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
-                return
-            }
-            viewModel.signIn(providerID: "apple.com", idToken: idTokenString, rawNonce: nonce)
-        }
+        viewModel.signIn(authorization: authorization)
     }
     
     func authorizationController(controller: ASAuthorizationController, didCompleteWithError error: Error) {
-        
         print("Sign in with Apple errored: \(error)")
     }
 }
