+++ DangProject/Presentation Layer/Login/ViewModel/LoginViewModel.swift
@@ -4,7 +4,8 @@
 //
 //  Created by 김성원 on 2022/04/27.
 //
-
+import AuthenticationServices
+import CryptoKit
 import Foundation
 
 import RxSwift
@@ -12,7 +13,7 @@ import RxRelay
 
 
 protocol LoginViewModelInput {
-    func signIn(providerID: String, idToken: String, rawNonce: String)
+    func signIn(authorization: ASAuthorization)
 }
 
 protocol LoginViewModelOutput {
@@ -61,6 +62,8 @@ class LoginViewModel: LoginViewModelProtocol {
     }
     
     //MARK: - Private Method
+    fileprivate var currentNonce: String?
+    
     private func checkProfileExistence(uid: String) {
         firebaseFireStoreUseCase.getProfileExistence(uid: uid)
     }
@@ -70,14 +73,82 @@ class LoginViewModel: LoginViewModelProtocol {
         userDefaults.set(uid, forKey: UserInfoKey.firebaseUID)
     }
     
+    private func randomNonceString(length: Int = 32) -> String {
+        precondition(length > 0)
+        let charset: [Character] =
+        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+        var result = ""
+        var remainingLength = length
+        
+        while remainingLength > 0 {
+            let randoms: [UInt8] = (0 ..< 16).map { _ in
+                var random: UInt8 = 0
+                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
+                if errorCode != errSecSuccess {
+                    fatalError(
+                        "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
+                    )
+                }
+                return random
+            }
+            randoms.forEach { random in
+                if remainingLength == 0 {
+                    return
+                }
+                if random < charset.count {
+                    result.append(charset[Int(random)])
+                    remainingLength -= 1
+                }
+            }
+        }
+        
+        return result
+    }
+    
+    @available(iOS 13, *)
+    private func sha256(_ input: String) -> String {
+        let inputData = Data(input.utf8)
+        let hashedData = SHA256.hash(data: inputData)
+        let hashString = hashedData.compactMap {
+            String(format: "%02x", $0)
+        }.joined()
+        
+        return hashString
+    }
+    
     //MARK: - Input
-    func signIn(providerID: String, idToken: String, rawNonce: String) {
-        firebaseAuthUseCase.requireFirebaseUID(providerID: providerID, idToken: idToken, rawNonce: rawNonce)
+    func loginButtonDidTap(with viewController: UIViewController) {
+        let nonce = randomNonceString()
+        currentNonce = nonce
+        let appleIDProvider = ASAuthorizationAppleIDProvider()
+        let request = appleIDProvider.createRequest()
+        request.requestedScopes = [.fullName, .email]
+        request.nonce = sha256(nonce)
+        
+        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
+        authorizationController.delegate = viewController as? ASAuthorizationControllerDelegate
+        authorizationController.presentationContextProvider = viewController as? ASAuthorizationControllerPresentationContextProviding
+        authorizationController.performRequests()
+    }
+    func signIn(authorization: ASAuthorization) {
+        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
+            guard let nonce = currentNonce else {
+                fatalError("Invalid state: A login callback was received, but no login request was sent.")
+            }
+            guard let appleIDToken = appleIDCredential.identityToken else {
+                print("Unable to fetch identity token")
+                return
+            }
+            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
+                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
+                return
+            }
+            firebaseAuthUseCase.requireFirebaseUID(providerID: "apple.com",
+                                                   idToken: idTokenString,
+                                                   rawNonce: nonce)
+        }
     }
    
     //MARK: - Output
     let profileExistenceObservable = PublishRelay<Bool>()
-    
-    
-    
 }
